% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neighboot-package.R
\docType{package}
\name{Neighboot-package}
\alias{Neighboot}
\alias{Neighboot-package}
\title{Neighboot: Neighborhood Bootstrap Method for RDS}
\description{
A bootstrap method for Respondent-Driven Sampling (RDS) that relies on the underlying structure of the RDS network to estimate uncertainty.
}
\keyword{internal}
