% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neighboot-package.R
\docType{data}
\name{pop.network}
\alias{pop.network}
\title{Population network}
\format{
A list containing two elements:
\describe{
\item{\code{traits}}{a dataframe of 2000 rows and 4 columns }
\item{\code{adj.mat}}{an adjacency matrix}
}
}
\usage{
pop.network
}
\description{
Population network
}
\keyword{datasets}
